+++ openssl-sys/src/lib.rs
@@ -24,10 +24,7 @@ pub use asn1::*;
 pub use bio::*;
 pub use bn::*;
 pub use cms::*;
-pub use conf::*;
 pub use crypto::*;
-pub use dh::*;
-pub use dsa::*;
 pub use dtls1::*;
 pub use ec::*;
 pub use err::*;
@@ -49,6 +46,7 @@ pub use srtp::*;
 pub use ssl::*;
 pub use ssl3::*;
 pub use tls1::*;
+pub use types::*;
 pub use x509::*;
 pub use x509_vfy::*;
 pub use x509v3::*;
@@ -61,10 +59,7 @@ mod asn1;
 mod bio;
 mod bn;
 mod cms;
-mod conf;
 mod crypto;
-mod dh;
-mod dsa;
 mod dtls1;
 mod ec;
 mod err;
@@ -86,6 +81,7 @@ mod srtp;
 mod ssl;
 mod ssl3;
 mod tls1;
+mod types;
 mod x509;
 mod x509_vfy;
 mod x509v3;
